backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.1
pydantic>=2.9.0
pydantic-settings>=2.5.0
playwright>=1.44.0
httpx>=0.27.0
python-dateutil>=2.9.0
python-dotenv>=1.0.1
